+++ src/graphql/schema/mod.rs
@@ -1,52 +1,47 @@
-use async_graphql::dynamic::{Object, Scalar, Schema, SchemaBuilder, TypeRef};
+use async_graphql::dynamic::{Object, Scalar, Schema, SchemaBuilder};
 use log::{debug, error, info};
+use serde::{Deserialize, Serialize};
 
 use crate::{configuration::subgraph::SubGraphConfig, data_sources::DataSources};
 
-mod generate_entities;
+mod create_entities;
 
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy, Debug, Deserialize, Serialize, PartialEq)]
 pub enum ResolverType {
     FindOne,
     FindMany,
     CreateOne,
-    // CreateMany,
-    // DeleteOne,
-    // DeleteMany,
-    // UpdateOne,
-    // UpdateMany,
+    UpdateOne,
 }
 
-#[derive(Debug)]
-pub struct ResolverConfig {
-    resolver_name: String,
-    return_type: TypeRef,
-}
-
-pub struct ServiceSchema {
+pub struct ServiceSchemaBuilder {
     pub subgraph_config: SubGraphConfig,
     pub schema_builder: SchemaBuilder,
     pub query: Object,
     pub mutation: Object,
+    pub data_sources: DataSources,
 }
 
-impl ServiceSchema {
-    pub fn build(subgraph_config: SubGraphConfig, data_sources: DataSources) -> Self {
-        ServiceSchema {
+impl ServiceSchemaBuilder {
+    pub fn new(subgraph_config: SubGraphConfig, data_sources: DataSources) -> Self {
+        info!("Creating Service Schema");
+        ServiceSchemaBuilder {
             subgraph_config,
             schema_builder: Schema::build("Query", Some("Mutation"), None)
-                .data(data_sources)
+                .data(data_sources.clone())
                 .enable_federation(),
             query: Object::new("Query").extends(),
             mutation: Object::new("Mutation"),
+            data_sources,
         }
     }
 
-    pub fn finish(mut self) -> Schema {
+    pub fn build(mut self) -> Schema {
+        info!("Building Schema");
+
         let object_id = Scalar::new("ObjectID");
 
-        self = self.generate_entities();
-        info!("Finishing Schema");
+        self = self.create_entities();
 
         let schema = self
             .schema_builder
@@ -54,16 +49,15 @@ impl ServiceSchema {
             .register(self.query)
             .register(self.mutation)
             .finish();
-        debug!("{:?}", schema);
 
-        let finished = match schema {
+        debug!("Schema Created: {:?}", schema);
+
+        match schema {
             Ok(sch) => sch,
             Err(err) => {
                 error!("{}", err);
                 panic!("Failed to build schema.")
             }
-        };
-
-        finished
+        }
     }
 }